+++ tests/test_file_system_cache.py
@@ -46,20 +46,12 @@ class CustomHashingMethodCache(FileSystemCache):
         super().__init__(*args, hash_method=hashlib.sha256, **kwargs)
 
 
-class CustomDefaultHashingMethodCache(FileSystemCache):
-    _default_hash_method = hashlib.sha256
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-
 @pytest.fixture(
     autouse=True,
     params=[
         FileSystemCache,
         CustomSerializerCache,
         CustomHashingMethodCache,
-        CustomDefaultHashingMethodCache,
     ],
 )
 def cache_factory(request, tmpdir):