+++ CHANGES.rst
@@ -3,6 +3,9 @@ Version 0.16.0
 
 unreleased
 
+- Change ``FileSystemCache`` default hash method to ``hashlib.sha256``
+  because ``hashlib.md5`` is not available in FIPS builds. :issue:`500`
+
 
 Version 0.15.4
 ---------------